refactor login banner for inner setting sections
